+++ src/errors/index.ts
@@ -1,6 +1,4 @@
 export * from "./error-codes.js";
 export * from "./contlify-error.js";
-export * from "./authentication-error.js";
-export * from "./validation-error.js";
 export * from "./not-found-error.js";
 export * from "./adapter-error.js";